+++ CHANGELOG.md
@@ -15,19 +15,27 @@ All notable changes to this project will be documented in this file.
 - Column "reference" on FormQuestion table of the database structure
 - Fields "reference" on forms and questions on oxe-web-admin
 
-## [1.13.2] - 2022-XX-XX
+## [1.13.2] - 2022-09-15
+
+### Added
+
+- Add form/extract_form resource on API
+- "Export" tab on Form item on oxe-web-admin
 
 ### Changed
 
 - Change from log to error dialog when failing to load a file
 - Dependent package versions
 - Recommended NodeJS version to 16.17.0
+- Review doc of entity/extract_entities
+- Server installation documentation
 
 ### Fixed
 
 - ISSUE#39 FIX: Change password reset box to show in green on success rather than in red
 - ISSUE#41 FIX: Claiming an entity requires logout/login to effect change in the profile
 - Vulnerabilities on dev packages dependencies
+- The answer is deleted when empty field on forms on oxe-web-community
 
 ## [1.13.1] - 2022-08-31